requirements.txt
pytesseract==0.3.11
pillow==10.0.0
opencv-python==4.8.0
pdf2image==1.16.3
spacy==3.8.0
tqdm==4.66.1
pandas==2.2.2
python-dotenv==1.0.0
